requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr_loader==0.2.7.post0
colorama==0.4.6
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-extensions==3.0.2
django-health-check==3.18.3
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
filelock==3.16.1
git-filter-repo==2.47.0
graphviz==0.14
gunicorn==23.0.0
idna==3.10
kombu==5.5.0
msgpack==1.1.0
mysql-connector-python==8.0.20
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pi==0.1.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==3.12.2
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydot==1.4.1
pyfingerprint==1.5
PyJWT==2.10.1
pyparsing==2.4.7
pyserial==3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pythonnet==3.0.5
pytz==2020.1
pyusb==1.3.1
pywin32==308
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.3.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
whitenoise==6.9.0
